+++ src/Data/Vector/Lens.hs
@@ -37,7 +37,6 @@ module Data.Vector.Lens
 
 import Control.Applicative
 import Control.Lens
-import Control.Lens.Internal
 import Data.Vector as Vector hiding (zip, filter)
 import Prelude hiding ((++), length, head, tail, init, last, map, reverse)
 import Data.List (nub)
@@ -118,7 +117,7 @@ forced = isos force force force force
 --
 -- @ordinal n@ is only a valid 'Lens' into a 'Vector' with 'length' at least @n + 1@.
 ordinal :: Int -> SimpleIndexedLens Int (Vector a) a
-ordinal i = index $ \ f v -> (\ a -> v // [(i, a)]) <$> f i (v ! i)
+ordinal i = indexing $ \ f v -> (\ a -> v // [(i, a)]) <$> f i (v ! i)
 {-# INLINE ordinal #-}
 
 -- | This 'Traversal' will ignore any duplicates in the supplied list of indices.
@@ -126,7 +125,7 @@ ordinal i = index $ \ f v -> (\ a -> v // [(i, a)]) <$> f i (v ! i)
 -- >>> toListOf (ordinals [1,3,2,5,9,10]) $ Vector.fromList [2,4..40]
 -- [4,8,6,12,20,22]
 ordinals :: [Int] -> SimpleIndexedTraversal Int (Vector a) a
-ordinals is = index $ \ f v -> let
+ordinals is = indexing $ \ f v -> let
      l = length v
      is' = nub $ filter (<l) is
   in fmap ((v //) . zip is') . traverse (uncurry f) . zip is $ fmap (v !) is'