+++ CHANGELOG.markdown
@@ -7,13 +7,14 @@
 * Removed `peruse`, `peruses`.
 * Generalized `review` and `reviews` to both return a `MonadReader` and to work on a `Projection`.
 * Added `reuse` : `use` :: `review` : `view` and `reuses` : `uses` :: `reviews` : `views` for working a `Projection` from the current `MonadState`.
-* Moved `Indexed(index)` to `Control.Lens.Internal`. `index` is a rather too common name to export by default.
 * Added `Control.Lens.Wrapped` providing a canonical isomorphism for newtypes.
 * Repurposed `ala` to be closer to the original design in @newtype@.
 * Added `au` and `alaf`.
 * Removed many isomorphisms for various newtypes. `_const`, `identity`, `_sum`, etc. Use `wrapping Const`, `wrapping Identity`, etc.
 * Removed `Data.Monoid.Lens` now that its newtypes are instances of `Wrapped`, exporting the (`<>=`)-variants from `Control.Lens.*`.
 * Renamed `via` to `cloneIso` for consistency.
+* Moved `Indexed(..)` to `Control.Lens.Classes`.
+* Renamed `index` to `indexing` to reduce conflicts with third-party libraries.
 
 3.6
 -----